requirements.txt
PyMuPDF>=1.24.0
openpyxl>=3.1.0
python-docx>=1.1.0
anthropic>=0.40.0
pydantic>=2.0.0
gradio>=5.0.0
httpx>=0.27.0
